introduce parseData function
