+++ src/App.js
@@ -8,6 +8,10 @@ import { Subheader } from "./components/Subheader.js";
 import { Chart } from "./components/Chart.js";
 import { Filter } from "./components/Filter.js";
 import { Button } from "./components/Button.js";
+import { parseData } from "./functions/parseData";
+import { handleChangeHelper } from "./functions/handleChangeHelper";
+import { handleClickHelper } from "./functions/handleClickHelper";
+
 import { csvJSON } from "./functions/csvJSON";
 import { formatImpressions } from "./functions/formatImpressions";
 import { extractDate } from "./functions/extractDate";
@@ -16,16 +20,16 @@ import { getColumns } from "./functions/getColumns";
 import { getMetrics } from "./functions/getMetrics";
 import { getDimensions } from "./functions/getDimensions";
 import { formatDate } from "./functions/formatDate";
-import { handleChangeHelper } from "./functions/handleChangeHelper";
-import { handleClickHelper } from "./functions/handleClickHelper";
 
 function App() {
+  const [isLoading, setIsLoading] = useState(true);
   const [data, setData] = useState([]);
   const [metrics, setMetrics] = useState([]);
   const [dimensions, setDimensions] = useState([]);
   const [selectedValues, setSelectedValues] = useState({});
   const [useFilters, setUseFilters] = useState(false);
   const [filteredData, setFilteredData] = useState(0);
+
   useEffect(() => {
     async function fetchData() {
       const response = await fetch(
@@ -33,18 +37,10 @@ function App() {
         //"https://raw.githubusercontent.com/addania/challenge/master/src/data/source.csv?raw=true"
         "http://adverity-challenge.s3-website-eu-west-1.amazonaws.com/DAMKBAoDBwoDBAkOBAYFCw.csv"
       );
-      const csvData = await response.text();
-      const jsonData = csvJSON(csvData);
-      const formattedImpressions = formatImpressions(jsonData);
-      const formattedDates = extractDate(formattedImpressions);
-      const sortedData = sortArray(formattedDates);
-      const tableColumns = getColumns(sortedData[0]);
-      const metricColumns = getMetrics(sortedData[0], tableColumns);
-      const dimensionColumns = getDimensions(sortedData[0], tableColumns);
-      const finalData = formatDate(sortedData);
-      setData(finalData);
-      setMetrics(metricColumns);
-      setDimensions(dimensionColumns);
+      const rawData = parseData(await response.text());
+      setData(rawData[0]);
+      setMetrics(rawData[1]);
+      setDimensions(rawData[2]);
     }
     fetchData();
   }, []);
@@ -70,7 +66,6 @@ function App() {
     setUseFilters(newState[0]);
     setFilteredData(newState[1]);
   }
-
   return (
     <div className="App" data-testid="app">
       <Container>
